refactor: move model download screen
